Move ScoreBoard to it's own file
