+++ src/shared/Common.js
@@ -1,6 +1,5 @@
-import React, { Component } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import pawCoinImg from './paw_coin.png';
 
 export const SceneWrapper = styled.main`
     background: #cfa7a7;
@@ -34,22 +33,3 @@ export const Backdrop = styled.div`
     position: absolute;
     z-index: 0;
 `;
-
-const PawCoin = styled.img.attrs({
-    src: pawCoinImg,
-    alt: 'one paw coin'
-})`
-    width: 4em;
-    height: 4em;
-`;
-
-const Purse = styled.div`
-    position: absolute;
-    top: 2vh;
-    right: 2vh;
-`;
-
-export const ScoreBoard = (props) =>
-    <Purse>
-        {props.progress.map((stage) => <PawCoin key={stage} />)}
-    </Purse>;
